Conflicts:
	README.textile

+++ spec/cobweb/crawl_worker_spec.rb
@@ -4,25 +4,28 @@ require File.expand_path(File.dirname(__FILE__) + '/../spec_helper')
 describe CrawlWorker, :local_only => true do
 
   before(:all) do
-    #store all existing resque process ids so we don't kill them afterwards
-    @existing_processes = `ps aux | grep sidekiq | grep -v grep | awk '{print $2}'`.split("\n")
-    puts @existing_processes
-    @existing_processes.should be_empty
-  
-    # START WORKERS ONLY FOR CRAWL QUEUE SO WE CAN COUNT ENQUEUED PROCESS AND FINISH QUEUES
-    puts "Starting Workers... Please Wait..."
-    `mkdir log`
-    `rm -rf output.log`
-    io = IO.popen("nohup sidekiq -r ./lib/crawl_worker.rb -q crawl_worker > ./log/output.log &")
-    puts "Workers Started."
-  
+    
+    if SIDEKIQ_INSTALLED    
+      #store all existing resque process ids so we don't kill them afterwards
+      @existing_processes = `ps aux | grep sidekiq | grep -v grep | awk '{print $2}'`.split("\n")
+
+      raise "Sidekiq is already running, please stop before running specs." if @existing_processes.count > 0
+    
+      # START WORKERS ONLY FOR CRAWL QUEUE SO WE CAN COUNT ENQUEUED PROCESS AND FINISH QUEUES
+      puts "Starting Workers... Please Wait..."
+      `mkdir log`
+      `rm -rf output.log`
+      io = IO.popen("nohup sidekiq -r ./lib/crawl_worker.rb -q crawl_worker > ./log/output.log &")
+      puts "Workers Started."
+    end  
   end
 
   before(:each) do
+    pending("Sidkiq not installed") unless SIDEKIQ_INSTALLED
     @base_url = "http://localhost:3532/"
     @base_page_count = 77
   
-    clear_queues
+    clear_sidekiq_queues
   end
 
   describe "with no crawl limit" do
@@ -31,7 +34,7 @@ describe CrawlWorker, :local_only => true do
         :crawl_id => Digest::SHA1.hexdigest("#{Time.now.to_i}.#{Time.now.usec}"),
         :crawl_limit => nil,
         :quiet => false,
-        :debug => false,
+        :debug => true,
         :cache => nil,
         :queue_system => :sidekiq
       }
@@ -57,6 +60,7 @@ describe CrawlWorker, :local_only => true do
         :crawl_id => Digest::SHA1.hexdigest("#{Time.now.to_i}.#{Time.now.usec}"),
         :quiet => true,
         :cache => nil,
+        :debug => true,
         :queue_system => :sidekiq,
         :valid_mime_types => ["text/html"]
       }
@@ -84,6 +88,7 @@ describe CrawlWorker, :local_only => true do
       @request = {
         :crawl_id => Digest::SHA1.hexdigest("#{Time.now.to_i}.#{Time.now.usec}"),
         :quiet => true,
+        :debug => true,
         :queue_system => :sidekiq,
         :cache => nil
       }
@@ -133,7 +138,6 @@ describe CrawlWorker, :local_only => true do
           wait_for_crawl_finished crawl[:crawl_id]
         
           mime_types = CrawlProcessWorker.queue_items(0, 200).map{|job| JSON.parse(job)["args"][0]["mime_type"]}
-          ap mime_types
           mime_types.select{|m| m=="text/html"}.count.should == 5
         end
       end
@@ -183,11 +187,11 @@ describe CrawlWorker, :local_only => true do
         wait_for_crawl_finished crawl[:crawl_id]
         CrawlFinishedWorker.queue_size.should == 1
       end      
-      it "should not crawl 100 pages" do
+      it "should not crawl more than 100 pages" do
         crawl = @cobweb.start(@base_url)
         @stat = Stats.new({:crawl_id => crawl[:crawl_id]})
         wait_for_crawl_finished crawl[:crawl_id]
-        CrawlProcessWorker.queue_size.should_not == 100
+        CrawlProcessWorker.queue_size.should_not > 100
       end      
     end    
   end
@@ -198,7 +202,7 @@ describe CrawlWorker, :local_only => true do
       command = "kill #{(@all_processes - @existing_processes).join(" ")}"
       IO.popen(command)
     end
-    clear_queues
+    clear_sidekiq_queues
   end
 
 end
@@ -234,14 +238,14 @@ def running?(crawl_id)
   result
 end
 
-def clear_queues
+def clear_sidekiq_queues
   Sidekiq.redis do |conn|
     conn.smembers("queues").each do |queue_name|
       conn.del("queue:#{queue_name}")
       conn.srem("queues", queue_name)
     end
   end
-  sleep 2
+  sleep 5
   
   CrawlProcessWorker.queue_size.should == 0
   CrawlFinishedWorker.queue_size.should == 0
